+++ src/main/java/core/repositorie/LikeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import core.model.Like;
  
-public interface LikeRepository extends JpaRepository<Like, Integer> {
+public interface LikeRepository extends JpaRepository<Like, Long> {
     
 }
